@@ -2,8 +2,6 @@ package de.caritas.cob.consultingtypeservice.api.controller;
 
 import static de.caritas.cob.consultingtypeservice.api.auth.UserRole.TENANT_ADMIN;
 import static de.caritas.cob.consultingtypeservice.api.auth.UserRole.TOPIC_ADMIN;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.authentication;
 import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
@@ -12,30 +10,20 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import de.caritas.cob.consultingtypeservice.ConsultingTypeServiceApplication;
-import de.caritas.cob.consultingtypeservice.api.auth.UserRole;
 import de.caritas.cob.consultingtypeservice.api.model.ApplicationSettingsEntity;
 import de.caritas.cob.consultingtypeservice.api.model.ApplicationSettingsPatchDTO;
 import de.caritas.cob.consultingtypeservice.api.repository.ApplicationSettingsRepository;
 import de.caritas.cob.consultingtypeservice.api.tenant.TenantContext;
 import de.caritas.cob.consultingtypeservice.api.util.JsonConverter;
-import java.util.Map;
-import javax.servlet.http.Cookie;
-import org.assertj.core.util.Lists;
-import org.assertj.core.util.Maps;
-import org.assertj.core.util.Sets;
+import de.caritas.cob.consultingtypeservice.testHelper.MongoTestInitializer;
+import jakarta.servlet.http.Cookie;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.keycloak.adapters.RefreshableKeycloakSecurityContext;
-import org.keycloak.adapters.spi.KeycloakAccount;
-import org.keycloak.adapters.springsecurity.account.SimpleKeycloakAccount;
-import org.keycloak.adapters.springsecurity.token.KeycloakAuthenticationToken;
-import org.keycloak.representations.AccessToken;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
@@ -43,6 +31,9 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import org.springframework.web.context.WebApplicationContext;
 
 @SpringBootTest(classes = ConsultingTypeServiceApplication.class)
+@ContextConfiguration(
+    classes = ConsultingTypeServiceApplication.class,
+    initializers = MongoTestInitializer.class)
 @TestPropertySource(properties = "spring.profiles.active=testing")
 @TestPropertySource(properties = "feature.multitenancy.with.single.domain.enabled=true")
 @AutoConfigureMockMvc(addFilters = false)
@@ -116,9 +107,9 @@ class ApplicationSettingsControllerIT {
                 .with(authentication(authentication))
                 .header("csrfHeader", "csrfToken")
                 .cookie(new Cookie("csrfCookie", "csrfToken"))
-                .contentType(javax.ws.rs.core.MediaType.APPLICATION_JSON)
+                .contentType(jakarta.ws.rs.core.MediaType.APPLICATION_JSON)
                 .content(jsonRequest)
-                .contentType(javax.ws.rs.core.MediaType.APPLICATION_JSON))
+                .contentType(jakarta.ws.rs.core.MediaType.APPLICATION_JSON))
         .andExpect(status().isOk())
         .andExpect(jsonPath("$.multitenancyWithSingleDomainEnabled.value").value(true))
         .andExpect(jsonPath("$.multitenancyWithSingleDomainEnabled.readOnly").value(true))
@@ -165,9 +156,9 @@ class ApplicationSettingsControllerIT {
                 .with(authentication(authentication))
                 .header("csrfHeader", "csrfToken")
                 .cookie(new Cookie("csrfCookie", "csrfToken"))
-                .contentType(javax.ws.rs.core.MediaType.APPLICATION_JSON)
+                .contentType(jakarta.ws.rs.core.MediaType.APPLICATION_JSON)
                 .content(jsonRequest)
-                .contentType(javax.ws.rs.core.MediaType.APPLICATION_JSON))
+                .contentType(jakarta.ws.rs.core.MediaType.APPLICATION_JSON))
         .andExpect(status().isOk());
   }
 
@@ -186,9 +177,9 @@ class ApplicationSettingsControllerIT {
                 .with(authentication(builder.withUserRole(TOPIC_ADMIN.getValue()).build()))
                 .header("csrfHeader", "csrfToken")
                 .cookie(new Cookie("csrfCookie", "csrfToken"))
-                .contentType(javax.ws.rs.core.MediaType.APPLICATION_JSON)
+                .contentType(jakarta.ws.rs.core.MediaType.APPLICATION_JSON)
                 .content(jsonRequest)
-                .contentType(javax.ws.rs.core.MediaType.APPLICATION_JSON))
+                .contentType(jakarta.ws.rs.core.MediaType.APPLICATION_JSON))
         .andExpect(status().isForbidden());
   }
 
@@ -197,23 +188,5 @@ class ApplicationSettingsControllerIT {
     entity.setReleaseToggles("featureToggleTenantCreationEnabled", true);
     applicationSettingsRepository.deleteAll();
     applicationSettingsRepository.save(entity);
-  }
-
-  private Authentication givenMockAuthentication(final UserRole authority) {
-    final var securityContext = mock(RefreshableKeycloakSecurityContext.class);
-    when(securityContext.getTokenString()).thenReturn("tokenString");
-    final var token = mock(AccessToken.class, Mockito.RETURNS_DEEP_STUBS);
-    when(securityContext.getToken()).thenReturn(token);
-    givenOtherClaimsAreDefinedForToken(token);
-    final KeycloakAccount mockAccount =
-        new SimpleKeycloakAccount(() -> "user", Sets.newHashSet(), securityContext);
-    return new KeycloakAuthenticationToken(
-        mockAccount, true, Lists.newArrayList((GrantedAuthority) authority::getValue));
-  }
-
-  private void givenOtherClaimsAreDefinedForToken(final AccessToken token) {
-    final Map<String, Object> claimMap = Maps.newHashMap("username", "test");
-    claimMap.put("userId", "some userid");
-    when(token.getOtherClaims()).thenReturn(claimMap);
   }
 }
