@@ -2,23 +2,28 @@ package de.caritas.cob.consultingtypeservice.config;
 
 import static java.util.Objects.nonNull;
 
+import com.google.common.collect.Lists;
 import de.caritas.cob.consultingtypeservice.api.auth.AuthenticatedUser;
 import de.caritas.cob.consultingtypeservice.api.exception.KeycloakException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
 import java.util.stream.Collectors;
-import javax.servlet.http.HttpServletRequest;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
 import lombok.Data;
 import org.hibernate.validator.constraints.URL;
 import org.keycloak.adapters.KeycloakConfigResolver;
 import org.keycloak.adapters.springboot.KeycloakSpringBootConfigResolver;
-import org.keycloak.adapters.springsecurity.token.KeycloakAuthenticationToken;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Scope;
 import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.context.WebApplicationContext;
 
@@ -30,28 +35,22 @@ public class KeycloakConfig {
 
   @Bean
   @Scope(scopeName = WebApplicationContext.SCOPE_REQUEST, proxyMode = ScopedProxyMode.TARGET_CLASS)
-  public KeycloakAuthenticationToken keycloakAuthenticationToken(HttpServletRequest request) {
-    return (KeycloakAuthenticationToken) request.getUserPrincipal();
-  }
-
-  @Bean
-  @Scope(scopeName = WebApplicationContext.SCOPE_REQUEST, proxyMode = ScopedProxyMode.TARGET_CLASS)
-  public AuthenticatedUser authenticatedUser(HttpServletRequest request) {
+  AuthenticatedUser authenticatedUser(HttpServletRequest request) {
     var userPrincipal = request.getUserPrincipal();
     var authenticatedUser = new AuthenticatedUser();
 
     if (nonNull(userPrincipal)) {
-      var authToken = (KeycloakAuthenticationToken) userPrincipal;
-      var securityContext = authToken.getAccount().getKeycloakSecurityContext();
-      var claimMap = securityContext.getToken().getOtherClaims();
+      var authToken = (JwtAuthenticationToken) userPrincipal;
+      var token = authToken.getToken(); // Extract the JWT token from the authentication token
+      var claimMap = token.getClaims(); // Retrieve claims from the JWT token
 
       try {
         if (claimMap.containsKey("username")) {
           authenticatedUser.setUsername(claimMap.get("username").toString());
         }
         authenticatedUser.setUserId(claimMap.get("userId").toString());
-        authenticatedUser.setAccessToken(securityContext.getTokenString());
-        authenticatedUser.setRoles(securityContext.getToken().getRealmAccess().getRoles());
+        authenticatedUser.setAccessToken(token.getTokenValue());
+        authenticatedUser.setRoles(extractRealmRoles(token).stream().collect(Collectors.toSet()));
       } catch (Exception exception) {
         throw new KeycloakException("Keycloak data missing.", exception);
       }
@@ -64,6 +63,17 @@ public class KeycloakConfig {
     }
 
     return authenticatedUser;
+  }
+
+  public Collection<String> extractRealmRoles(Jwt jwt) {
+    Map<String, Object> realmAccess = (Map<String, Object>) jwt.getClaims().get("realm_access");
+    if (realmAccess != null) {
+      var roles = (List<String>) realmAccess.get("roles");
+      if (roles != null) {
+        return roles;
+      }
+    }
+    return Lists.newArrayList();
   }
 
   @Bean
