@@ -11,22 +11,27 @@ import de.caritas.cob.consultingtypeservice.api.admin.service.ConsultingTypeAdmi
 import de.caritas.cob.consultingtypeservice.api.consultingtypes.ConsultingTypeMongoRepositoryService;
 import de.caritas.cob.consultingtypeservice.api.model.ExtendedConsultingTypeResponseDTO;
 import de.caritas.cob.consultingtypeservice.api.model.PaginationLinks;
+import de.caritas.cob.consultingtypeservice.testHelper.MongoTestInitializer;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 
-@SpringBootTest(classes = ConsultingTypeServiceApplication.class)
+@SpringBootTest
+@ContextConfiguration(
+    classes = ConsultingTypeServiceApplication.class,
+    initializers = MongoTestInitializer.class)
 @TestPropertySource(properties = "spring.profiles.active=testing")
-public class ConsultingTypeAdminControllerIT {
+class ConsultingTypeAdminControllerIT {
 
   @Autowired private ConsultingTypeAdminService consultingTypeAdminService;
 
   @Autowired private ConsultingTypeMongoRepositoryService consultingTypeMongoRepositoryService;
 
   @Test
-  public void findConsultingTypes_Should_returnOneResult_When_perPageIsSetToOneAndPageIsSetToOne() {
+  void findConsultingTypes_Should_returnOneResult_When_perPageIsSetToOneAndPageIsSetToOne() {
 
     List<ExtendedConsultingTypeResponseDTO> consultingTypes =
         this.consultingTypeAdminService.findConsultingTypes(1, 1).getEmbedded();
@@ -35,7 +40,7 @@ public class ConsultingTypeAdminControllerIT {
   }
 
   @Test
-  public void findConsultingTypes_Should_returnOneResult_When_paginationParamsAreZero() {
+  void findConsultingTypes_Should_returnOneResult_When_paginationParamsAreZero() {
     List<ExtendedConsultingTypeResponseDTO> consultingTypes =
         this.consultingTypeAdminService.findConsultingTypes(0, 0).getEmbedded();
 
@@ -43,7 +48,7 @@ public class ConsultingTypeAdminControllerIT {
   }
 
   @Test
-  public void findConsultingTypes_Should_returnOneResult_When_paginationParamsAreNegative() {
+  void findConsultingTypes_Should_returnOneResult_When_paginationParamsAreNegative() {
     List<ExtendedConsultingTypeResponseDTO> consultingTypes =
         this.consultingTypeAdminService.findConsultingTypes(-100, -1000).getEmbedded();
 
@@ -51,8 +56,7 @@ public class ConsultingTypeAdminControllerIT {
   }
 
   @Test
-  public void
-      findConsultingTypes_Should_returnPaginatedEntities_When_paginationParamsAreSplitted() {
+  void findConsultingTypes_Should_returnPaginatedEntities_When_paginationParamsAreSplitted() {
     List<ExtendedConsultingTypeResponseDTO> firstPage =
         this.consultingTypeAdminService.findConsultingTypes(0, 2).getEmbedded();
     List<ExtendedConsultingTypeResponseDTO> secondPage =
@@ -63,7 +67,7 @@ public class ConsultingTypeAdminControllerIT {
   }
 
   @Test
-  public void findConsultingTypes_Should_haveExpectedLinks_When_AllParamsAreProvided() {
+  void findConsultingTypes_Should_haveExpectedLinks_When_AllParamsAreProvided() {
     PaginationLinks paginationLinks =
         this.consultingTypeAdminService.findConsultingTypes(1, 2).getLinks();
 
@@ -79,8 +83,7 @@ public class ConsultingTypeAdminControllerIT {
   }
 
   @Test
-  public void
-      findConsultingTypes_Should_returnAllConsultingTypes_When_ProvidedWithMaxPerPagesParam() {
+  void findConsultingTypes_Should_returnAllConsultingTypes_When_ProvidedWithMaxPerPagesParam() {
     List<ExtendedConsultingTypeResponseDTO> page =
         this.consultingTypeAdminService.findConsultingTypes(0, Integer.MAX_VALUE).getEmbedded();
 
