@@ -10,21 +10,26 @@ import de.caritas.cob.consultingtypeservice.ConsultingTypeServiceApplication;
 import de.caritas.cob.consultingtypeservice.api.consultingtypes.ConsultingTypeRepositoryService;
 import de.caritas.cob.consultingtypeservice.api.model.ExtendedConsultingTypeResponseDTO;
 import de.caritas.cob.consultingtypeservice.api.model.PaginationLinks;
+import de.caritas.cob.consultingtypeservice.testHelper.MongoTestInitializer;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 
-@SpringBootTest(classes = ConsultingTypeServiceApplication.class)
+@SpringBootTest
+@ContextConfiguration(
+    classes = ConsultingTypeServiceApplication.class,
+    initializers = MongoTestInitializer.class)
 @TestPropertySource(properties = "spring.profiles.active=testing")
-public class ConsultingTypeAdminServiceIT {
+class ConsultingTypeAdminServiceIT {
 
   @Autowired private ConsultingTypeAdminService consultingTypeAdminService;
   @Autowired private ConsultingTypeRepositoryService consultingTypeRepositoryService;
 
   @Test
-  public void findConsultingTypes_Should_returnOneResult_When_perPageIsSetToOneAndPageIsSetToOne() {
+  void findConsultingTypes_Should_returnOneResult_When_perPageIsSetToOneAndPageIsSetToOne() {
 
     List<ExtendedConsultingTypeResponseDTO> consultingTypes =
         this.consultingTypeAdminService.findConsultingTypes(1, 1).getEmbedded();
@@ -33,7 +38,7 @@ public class ConsultingTypeAdminServiceIT {
   }
 
   @Test
-  public void findConsultingTypes_Should_returnOneResult_When_paginationParamsAreZero() {
+  void findConsultingTypes_Should_returnOneResult_When_paginationParamsAreZero() {
     List<ExtendedConsultingTypeResponseDTO> consultingTypes =
         this.consultingTypeAdminService.findConsultingTypes(0, 0).getEmbedded();
 
@@ -41,7 +46,7 @@ public class ConsultingTypeAdminServiceIT {
   }
 
   @Test
-  public void findConsultingTypes_Should_returnOneResult_When_paginationParamsAreNegative() {
+  void findConsultingTypes_Should_returnOneResult_When_paginationParamsAreNegative() {
     List<ExtendedConsultingTypeResponseDTO> consultingTypes =
         this.consultingTypeAdminService.findConsultingTypes(-100, -1000).getEmbedded();
 
@@ -49,8 +54,7 @@ public class ConsultingTypeAdminServiceIT {
   }
 
   @Test
-  public void
-      findConsultingTypes_Should_returnPaginatedEntities_When_paginationParamsAreSplitted() {
+  void findConsultingTypes_Should_returnPaginatedEntities_When_paginationParamsAreSplitted() {
     List<ExtendedConsultingTypeResponseDTO> firstPage =
         this.consultingTypeAdminService.findConsultingTypes(0, 4).getEmbedded();
     List<ExtendedConsultingTypeResponseDTO> secondPage =
@@ -61,7 +65,7 @@ public class ConsultingTypeAdminServiceIT {
   }
 
   @Test
-  public void findConsultingTypes_Should_haveExpectedLinks_When_AllParamsAreProvided() {
+  void findConsultingTypes_Should_haveExpectedLinks_When_AllParamsAreProvided() {
     PaginationLinks paginationLinks =
         this.consultingTypeAdminService.findConsultingTypes(1, 1).getLinks();
 
@@ -77,8 +81,7 @@ public class ConsultingTypeAdminServiceIT {
   }
 
   @Test
-  public void
-      findConsultingTypes_Should_returnAllConsultingTypes_When_ProvidedWithMaxPerPagesParam() {
+  void findConsultingTypes_Should_returnAllConsultingTypes_When_ProvidedWithMaxPerPagesParam() {
     List<ExtendedConsultingTypeResponseDTO> page =
         this.consultingTypeAdminService.findConsultingTypes(0, Integer.MAX_VALUE).getEmbedded();
 
