@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import de.caritas.cob.consultingtypeservice.api.tenant.TenantContext;
 import de.caritas.cob.consultingtypeservice.schemas.model.ConsultingType;
+import de.caritas.cob.consultingtypeservice.testHelper.MongoTestInitializer;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.AfterEach;
@@ -11,26 +12,28 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 
 @SpringBootTest
-@AutoConfigureMockMvc(addFilters = false)
+@AutoConfigureMockMvc
 @TestPropertySource(properties = "spring.profiles.active=testing")
 @TestPropertySource(properties = "multitenancy.enabled=true")
+@ContextConfiguration(initializers = MongoTestInitializer.class)
 @TestPropertySource(
     properties =
         "consulting.types.json.path=src/test/resources/consulting-type-settings-tenant-specific")
-public class ConsultingTypeGroupRespositoryTenantAwareIT {
+class ConsultingTypeGroupRespositoryTenantAwareIT {
 
   @Autowired private ConsultingTypeGroupRepository consultingTypeGroupRepository;
 
   @AfterEach
-  public void tearDown() {
+  public void teardown() {
     TenantContext.clear();
   }
 
   @Test
-  public void getConsultingTypesGroupMap_Should_ReturnMapWithConsultingTypeGroups() {
+  void getConsultingTypesGroupMap_Should_ReturnMapWithConsultingTypeGroups() {
     // given
     TenantContext.setCurrentTenant(2L);
 
@@ -49,7 +52,6 @@ public class ConsultingTypeGroupRespositoryTenantAwareIT {
     final String GROUP_2 = "group2";
     final String GROUP_3 = "group3";
     final int CONSULTING_TYPE_ID_1 = 11;
-    final int CONSULTING_TYPE_ID_2 = 12;
     assertThat(result.get(GROUP_2)).hasSize(1);
     assertThat(result.get(GROUP_3)).hasSize(1);
     assertThat(result).containsKey(GROUP_2);
