@@ -8,6 +8,7 @@ import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import de.caritas.cob.consultingtypeservice.ConsultingTypeServiceApplication;
 import de.caritas.cob.consultingtypeservice.api.model.ConsultingTypeEntity;
 import de.caritas.cob.consultingtypeservice.schemas.model.ConsultingType;
+import de.caritas.cob.consultingtypeservice.testHelper.MongoTestInitializer;
 import java.io.IOException;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
@@ -19,23 +20,25 @@ import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
 
-@DataMongoTest()
-@ContextConfiguration(classes = ConsultingTypeServiceApplication.class)
+@DataMongoTest
+@ContextConfiguration(
+    classes = ConsultingTypeServiceApplication.class,
+    initializers = MongoTestInitializer.class)
 @TestPropertySource(properties = "spring.profiles.active=testing")
 @TestPropertySource(properties = "multitenancy.enabled=true")
 @TestPropertySource(
     properties =
         "consulting.types.json.path=src/test/resources/consulting-type-settings-tenant-specific")
-public class ConsultingTypeMongoRespositoryIT {
+class ConsultingTypeMongoRespositoryIT {
 
-  private final String MONGO_COLLECTION_NAME = "consulting_types";
+  private static final String MONGO_COLLECTION_NAME = "consulting_types";
 
   @Autowired private ConsultingTypeRepository consultingTypeRepository;
 
   @Autowired MongoTemplate mongoTemplate;
 
   @BeforeEach
-  public void initializeMongoDbWithData() throws IOException {
+  void initializeMongoDbWithData() throws IOException {
     mongoTemplate.dropCollection(MONGO_COLLECTION_NAME);
     insertJsonFromFilename("consulting-type-0.json");
     insertJsonFromFilename("consulting-type-1.json");
@@ -53,7 +56,7 @@ public class ConsultingTypeMongoRespositoryIT {
   }
 
   @Test
-  public void findByConsultingTypeId_Should_ReturnCorrectConsultingType() {
+  void findByConsultingTypeId_Should_ReturnCorrectConsultingType() {
     // given
     Integer consultingTypeId = 10;
     String slug = "consultingtype10";
@@ -67,7 +70,7 @@ public class ConsultingTypeMongoRespositoryIT {
   }
 
   @Test
-  public void findBySlug_Should_ReturnCorrectConsultingTyp() {
+  void findBySlug_Should_ReturnCorrectConsultingTyp() {
     // given
     Integer consultingTypeId = 10;
     String slug = "consultingtype10";
@@ -81,7 +84,7 @@ public class ConsultingTypeMongoRespositoryIT {
   }
 
   @Test
-  public void findAll_Should_ReturnAllConsultingTypes() {
+  void findAll_Should_ReturnAllConsultingTypes() {
     // when
     List<ConsultingTypeEntity> result = consultingTypeRepository.findAll();
     // then
