@@ -13,10 +13,10 @@ import java.util.List;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.junit.jupiter.api.Test;
 
-public class ConsultingTypeGroupResponseMapperTest {
+class ConsultingTypeGroupResponseMapperTest {
 
   @Test
-  public void mapConsultingType_Should_MapAllProperties() throws IOException {
+  void mapConsultingType_Should_MapAllProperties() throws IOException {
 
     List<ConsultingType> consultingTypeList = HelperMethods.getConsultingTypeList();
     ConsultingTypeGroupResponseDTO result =
